implement conflicts in datastore; split api
