+++ brewblox_devcon_spark/__main__.py
@@ -2,8 +2,10 @@
 Example of how to import and use the brewblox service
 """
 
-from brewblox_devcon_spark import (api, broadcaster, commander, commander_sim,
+from brewblox_devcon_spark import (broadcaster, commander, commander_sim,
                                    datastore, device)
+from brewblox_devcon_spark.api import (alias_api, conflict_api, debug_api,
+                                       object_api, profile_api, system_api)
 from brewblox_service import brewblox_logger, events, service
 
 LOGGER = brewblox_logger(__name__)
@@ -45,7 +47,14 @@ def main():
 
     datastore.setup(app)
     device.setup(app)
-    api.setup(app)
+
+    debug_api.setup(app)
+    alias_api.setup(app)
+    conflict_api.setup(app)
+    object_api.setup(app)
+    profile_api.setup(app)
+    system_api.setup(app)
+
     events.setup(app)
     broadcaster.setup(app)
