+++ brewblox_devcon_spark/api/__init__.py
@@ -1,0 +1,8 @@
+"""
+Namespace for all REST API modules.
+"""
+
+API_ID_KEY = 'id'
+API_TYPE_KEY = 'type'
+API_DATA_KEY = 'data'
+API_LIST_KEY = 'objects'
